backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
google-generativeai==0.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
matplotlib==3.8.0
pandas==2.1.1
scikit-learn==1.3.2
plotly==5.18.0
pymongo==4.6.1
motor==3.3.2 